Attempt to work with all of 3 LED digits
Some functional added.
Now it implied to display all of 3 digits of the LED 7-segment indicator, simply what stored in R21, R22 and R23 registers.
The trouble is that segments' leads are common for all the digits. So it needs a kind of advanced design in order to display different signs on different digits (what can't be haywired should be hardcoded, lol).
Well, at least the time of display delaying was adjusted great (150 Hz), LEDs are shining continuously.