Added index branching
I replaced a large set of CPI...BREQ with the procedure of Index JuMPs, and addresses storage in flash memory.
Tried to trace it through on the paper sheet, was failed. But somehow it works on chip, lol. It needs tests though.
Well, the most difficult thing for me is dealing with different addresses notations. The compiler does it in words, the MCU - in bytes. Well, soon or late I'll work it out by myself, somehow.